Remove incorrectly specified @UserIdInt
This is a follow up CL to my previous CL [1], which introduced
TextServicesManagerServicve.SessionRequest for internal state
management.

It turns out that SessionRequest#mUserId was incorrectly marked with
@UserIdInt, even though it actually stores uid, not user ID int.

This CL addresses the confusion by removing @UserIdInt as well as
renaming it to SessionRequest#mUid.

There should be no behavior change actually.

 [1]: I92e7aa40dc9ea14f67d355f0bfa15325b775d27b

Bug: 7254002
Test: compile
Change-Id: I94f19d028be3b24fe0f3d848236ea9ca68e537b4