Display that MAC is randomized in About Phones
When Connected MAC Randomization is enabled and the device is
disconnected, there isn't a valid MAC address to display since we don't
know what the device's next MAC address is going to be.

In such cases, display "MAC address is randomized" instead of an invalid
MAC address in About Phones page.

Bug: 73663275
Test: make RunSettingsLibRoboTests -j40
Change-Id: I0a4df2b1b0823632b26f90e097c26f47dd5a0a67