Clear pattern on completion (if fade is enabled).
This CL clears the pattern immediately once it's completed. This
ensures that a half-drawn pattern doesn't remain as an artifact on the
lock-screen while we display an error message.

This behavior is gated on pattern fading, so if that's disabled, then
we retain the previous behavior (of showing the incorrect pattern
along with the error message).

Test: Draw an incorrect pattern.
Bug: 74531517
Change-Id: I41569e3fe10f91b787ef22600beb66d44f98cd90